Prepare for publish
